@@ -3,147 +3,37 @@ package com.Shiva.SCM.entity;
 import java.util.ArrayList;
 import java.util.List;
 
-import javax.persistence.CascadeType;
-import javax.persistence.Column;
-import javax.persistence.Entity;
-import javax.persistence.FetchType;
-import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
-import javax.persistence.Id;
-import javax.persistence.OneToMany;
-import javax.persistence.Table;
-import javax.validation.constraints.NotBlank;
-import javax.validation.constraints.Size;
+import jakarta.persistence.*;
+import jakarta.validation.constraints.NotBlank;
+import jakarta.validation.constraints.Size;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
-
-
-
-
+@Data
+@NoArgsConstructor
 @Entity
-@Table(name="USER")
+@Table(name="USERS")
 public class User {
-	 @Id
-	 @GeneratedValue(strategy=GenerationType.AUTO)
-	private int id;
-	 @NotBlank(message="Name field is required !!")
-	 @Size(min=2,max=20,message="min 2 and max 20 charaecter are allowed !!")
-	private String name;
-	@Column(unique=true)
-	private String email;
-	private String password;
-	private String role;
-	private boolean enabled;
-	private String imageUrl;
-	@Column(length=500)
-	private String  about;
-	
-	@OneToMany(cascade=CascadeType.ALL,fetch=FetchType.LAZY,mappedBy="user",orphanRemoval=true)
-	private List<Contact> contacts=new ArrayList<>();
-	
-	
-	public User() {
-		super();
-		// TODO Auto-generated constructor stub
-	}
 
+    @Id
+    @GeneratedValue(strategy=GenerationType.AUTO)
+    private Integer id;   // ✅ int → Integer
 
-	public int getId() {
-		return id;
-	}
+    @NotBlank(message="Name field is required !!")
+    @Size(min=2,max=20,message="min 2 and max 20 character are allowed !!")
+    private String name;
 
+    @Column(unique=true)
+    private String email;
 
-	public void setId(int id) {
-		this.id = id;
-	}
+    private String password;
+    private String role;
+    private boolean enabled;
+    private String imageUrl;
 
+    @Column(length=500)
+    private String about;
 
-	public String getName() {
-		return name;
-	}
-
-
-	public void setName(String name) {
-		this.name = name;
-	}
-
-
-	public String getEmail() {
-		return email;
-	}
-
-
-	public void setEmail(String email) {
-		this.email = email;
-	}
-
-
-	public String getPassword() {
-		return password;
-	}
-
-
-	public void setPassword(String password) {
-		this.password = password;
-	}
-
-
-	public String getRole() {
-		return role;
-	}
-
-
-	public void setRole(String role) {
-		this.role = role;
-	}
-
-
-	public boolean isEnabled() {
-		return enabled;
-	}
-
-
-	public void setEnabled(boolean enabled) {
-		this.enabled = enabled;
-	}
-
-
-	public String getImageUrl() {
-		return imageUrl;
-	}
-
-
-	public void setImageUrl(String imageUrl) {
-		this.imageUrl = imageUrl;
-	}
-
-
-	public String getAbout() {
-		return about;
-	}
-
-
-	public void setAbout(String about) {
-		this.about = about;
-	}
-
-
-	public List<Contact> getContacts() {
-		return contacts;
-	}
-
-
-	public void setContacts(List<Contact> contacts) {
-		this.contacts = contacts;
-	}
-
-
-	@Override
-	public String toString() {
-		return "User [id=" + id + ", name=" + name + ", email=" + email + ", password=" + password + ", role=" + role
-				+ ", enabled=" + enabled + ", imageUrl=" + imageUrl + ", about=" + about + ", contacts=" + contacts
-				+ "]";
-	}
-	
-	
-
+    @OneToMany(cascade=CascadeType.ALL, fetch=FetchType.LAZY, mappedBy="user", orphanRemoval=true)
+    private List<Contact> contacts = new ArrayList<>();
 }
