@@ -2,67 +2,53 @@ package com.Shiva.SCM.config;
 
 import java.util.Collection;
 import java.util.List;
-
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
-
 import com.Shiva.SCM.entity.User;
+import lombok.Getter;
 
+@Getter
 public class CustomUserDetails implements UserDetails {
 
-	private User user;
+    private final User user;
 
-	public CustomUserDetails(User user) {
-		super();
-		this.user = user;
-	}
+    public CustomUserDetails(User user) {
+        this.user = user;
+    }
 
-	@Override
-	public Collection<? extends GrantedAuthority> getAuthorities() {
-		
-		SimpleGrantedAuthority simpleGrantedAuthority=new SimpleGrantedAuthority(user.getRole());
-				
-		
-		
-		
-		return List.of(simpleGrantedAuthority);
-	}
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return List.of(new SimpleGrantedAuthority(user.getRole()));
+    }
 
-	@Override
-	public String getPassword() {
+    @Override
+    public String getPassword() {
+        return user.getPassword();
+    }
 
-		return user.getPassword();
-	}
+    @Override
+    public String getUsername() {
+        return user.getEmail();
+    }
 
-	@Override
-	public String getUsername() {
+    @Override
+    public boolean isAccountNonExpired() {
+        return true;
+    }
 
-		return user.getEmail();
-	}
+    @Override
+    public boolean isAccountNonLocked() {
+        return true;
+    }
 
-	@Override
-	public boolean isAccountNonExpired() {
+    @Override
+    public boolean isCredentialsNonExpired() {
+        return true;
+    }
 
-		return true;
-	}
-
-	@Override
-	public boolean isAccountNonLocked() {
-
-		return true;
-	}
-
-	@Override
-	public boolean isCredentialsNonExpired() {
-
-		return true;
-	}
-
-	@Override
-	public boolean isEnabled() {
-
-		return true;
-	}
-
+    @Override
+    public boolean isEnabled() {
+        return user.isEnabled(); // ab user ki enabled field se sync hai
+    }
 }
