@@ -1,71 +1,21 @@
 package com.Shiva.SCM.config;
 
+import com.Shiva.SCM.dao.UserRepository;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
 import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 
 @Configuration
-@EnableWebSecurity
-public class MyConfig extends WebSecurityConfigurerAdapter{
-	
-	@Bean
-	public UserDetailsService getUserDetailService()
-	{
-		return new UserDetailsServiceImpl();
-	}
+public class MyConfig {
 
-	
-	
-	@Bean
-	public BCryptPasswordEncoder passwordEncoder()
-	{ 
-		
-		
-		return new  BCryptPasswordEncoder();
-	}
-	
-	 @Bean
-	public DaoAuthenticationProvider authenticationProvider()
-	{
-		
-		DaoAuthenticationProvider daoAuthenticationProvider=new DaoAuthenticationProvider();
-		
-		daoAuthenticationProvider.setUserDetailsService(getUserDetailService());
-		
-		daoAuthenticationProvider.setPasswordEncoder(passwordEncoder());
-		
-		return daoAuthenticationProvider;
-	}
+    private final UserRepository userRepository;
 
-     //configure method 
+    public MyConfig(UserRepository userRepository) {
+        this.userRepository = userRepository;
+    }
 
-	@Override
-	protected void configure(AuthenticationManagerBuilder auth) throws Exception {
-		auth.authenticationProvider(authenticationProvider());
-	
-	}
-
-
-
-	@Override
-	protected void configure(HttpSecurity http) throws Exception {
-		
-		http.authorizeHttpRequests().antMatchers("/admin/**").hasRole("ADMIN")
-		.antMatchers("/user/**").hasRole("USER")
-		.antMatchers("/**").permitAll().and().formLogin()
-		.loginPage("/signin")
-		.loginProcessingUrl("/dologin")// where want to submit
-		.defaultSuccessUrl("/user/index")// success hone k bat ye chalega
-		.and().csrf().disable();
-	}
-	
-	
-	 
-	 
+    @Bean
+    public UserDetailsService getUserDetailsService() {
+        return new UserDetailsServiceImpl(userRepository); // ✅ ab sahi
+    }
 }
