@@ -1,35 +1,28 @@
 package com.Shiva.SCM.config;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import com.Shiva.SCM.dao.UserRepository;
+import com.Shiva.SCM.entity.User;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
+import org.springframework.stereotype.Service;
 
-import com.Shiva.SCM.dao.UserRepository;
-import com.Shiva.SCM.entity.User;
+import lombok.RequiredArgsConstructor;
 
-public class UserDetailsServiceImpl implements UserDetailsService{
+@Service
+@RequiredArgsConstructor
+public class UserDetailsServiceImpl implements UserDetailsService {
 
-	
-	
-	  @Autowired
-	private UserRepository userRepository;
-	
-	
-	@Override
-	public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-		
-		// fetching user from database 
-      User user=userRepository.getUserByUserName(username);
-		
-		if(user==null)
-		{
-			throw new UsernameNotFoundException("Could not found user !!");
-		}
-		
-		CustomUserDetails customUserDetails=new CustomUserDetails(user);
-		
-		return customUserDetails;
-	}
+    private final UserRepository userRepository;
 
+    @Override
+    public UserDetails loadUserByUsername(String email) throws UsernameNotFoundException {
+        User user = userRepository.findByEmail(email);
+
+        if (user == null) {
+            throw new UsernameNotFoundException("Could not find user with email: " + email);
+        }
+
+        return new CustomUserDetails(user);
+    }
 }
