@@ -1,14 +1,17 @@
 package com.Shiva.SCM.dao;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-import org.springframework.data.repository.query.Param;
-
+import org.springframework.stereotype.Repository;
 import com.Shiva.SCM.entity.User;
 
-public interface UserRepository  extends JpaRepository<User,Integer>{
-	 
-	@Query("select u from User u where u.email=:email")
-	public User getUserByUserName(@Param("email") String email);
+@Repository
+public interface UserRepository extends JpaRepository<User, Integer> {
 
+    // Spring Data JPA convention-based method
+    User findByEmail(String email);
+
+    // Custom alias (same kaam karega)
+    default User getUserByUserName(String email) {
+        return findByEmail(email);
+    }
 }
