@@ -1,28 +1,13 @@
 package com.Shiva.SCM.helper;
 
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
 public class Message {
-   private String content;
-   private String type;
-public Message(String content, String type) {
-	super();
-	this.content = content;
-	this.type = type;
-}
-public Message() {
-	super();
-	// TODO Auto-generated constructor stub
-}
-public String getContent() {
-	return content;
-}
-public void setContent(String content) {
-	this.content = content;
-}
-public String getType() {
-	return type;
-}
-public void setType(String type) {
-	this.type = type;
-}
-   
+    private String content;
+    private String type;
 }
